add gpl2 statments in my code

+++ src/main.rs
@@ -1,3 +1,24 @@
+/*
+KNOW ALL MEN BY THESE PRESENTS: 'i': man [Dakota James Owen Keeler]
+copyright this software in the year of our lord 2017 under the GNU
+Public License version 2.
+Contact: bearzrobotics@gmail.com
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 2 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+live honorably, harm no one, give to each his own.
+*/
 extern crate clap;
 
 use clap::{Arg, App, AppSettings};
@@ -17,22 +38,26 @@ fn main() {
 			.long("debug")
 			.takes_value(false)
 			.help("This sets the debuing flag to true, printing out all debuging info.\n Don't run with verbose. This will print out all info that verbose does."))
-        .arg(Arg::with_name("config")
-            .short("c")
-            .long("config")
-            .value_name("FILE")
-            .help("Sets a custom config file")
-			.takes_value(true))
+		.arg(Arg::with_name("delete_all_cache")
+            .short("a")
+            .long("delete_allr")
+            .help("Deletes both user and system cache <Must be root>")
+			.takes_value(false))
 		.arg(Arg::with_name("delete_system_cache")
-            .short("D")
+            .short("s")
             .long("delete_system")
             .help("Delete System cache, Must be root")
-			.takes_value(true))
+			.takes_value(false))
+		.arg(Arg::with_name("delete_user_cache")
+            .short("u")
+            .long("delete_user")
+            .help("Delete user cache, If not running as root it will only delete the current users. \r\n <If no options supplied this is the defualt>")
+			.takes_value(false))
 		.arg(Arg::with_name("verbose")
 			.short("v")
 			.long("verbose")
 			.takes_value(false)
-			.help("This sets the verbose flag to true, printing out verbose info. Less than debug though.\n Don't run with debug"))
+			.help("This sets the verbose flag to true, printing out verbose info. Less than debug though.\r\n Don't run with debug"))
 		.get_matches();		
 
 	//let all_flag = matches.value_of("all");
@@ -67,7 +92,7 @@ fn main() {
 		}else {
 			println!("Not running as root");
 		}
-	}else{
+	}else if  matches.is_present("delete_all_cache"){
 		utils::create_log_file(0);
 		cleaner::delete_user_cache(0);
 
@@ -77,6 +102,10 @@ fn main() {
 		}else {
 			println!("Not running as root");
 		}
+	}else{
+		// By Defualt deletes user cache.
+		utils::create_log_file(0);
+		cleaner::delete_user_cache(0);
 	}		
 }
